Refactored pages to have authentication and rendering in a stateless higher-order component named PageWrapper.
