+++ pages/index.js
@@ -1,40 +1,13 @@
 import React from 'react';
 
-import Banner from '../components/stateless/Banner';
-import GoogleContainer from '../components/GoogleContainer';
-import BootstrapLink from 
-'../components/stateless/BootstrapLink';
+import PageWrapper from '../components/stateless/PageWrapper';
 
-import cookies from 'next-cookies';
-
-export default class extends React.Component {
-  static async getInitialProps(ctx) {
-    const { userEmail } = cookies(ctx);
-    if (userEmail !== undefined) {
-      return {
-        ...cookies(ctx),
-        authUrl: 'http://localhost:4000/auth/logout',
-        authText: 'Logout'
-      }
-    } else {
-      ctx.res.redirect('/login');
-    }
-  }
+class Index extends React.Component {
 
   render() {
-      return(
-
-        <div>
-          <BootstrapLink/>
-          <Banner name={this.props.firstName} authUrl={this.props.authUrl} authText={this.props.authText} />
-          <GoogleContainer
-            firstName = {this.props.firstName}
-            lastName = {this.props.lastName}
-            batch = {this.props.userBatches}
-            email = {this.props.userEmail}
-            isAdd = {false}
-          />
-        </div>
-      );
+      return(this.props.renderPage(false));
   }
 }
+
+
+export default PageWrapper(Index);
